- moves java json serialization into generic package name for consistency
