+++ serialization/java/json/lib/src/main/java/com/microsoft/kiota/serialization/JsonParseNode.java
@@ -1,4 +1,4 @@
-package com.microsoft.kiota.serialization.json;
+package com.microsoft.kiota.serialization;
 
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
