+++ serialization/java/json/lib/src/main/java/com/microsoft/kiota/serialization/JsonSerializationWriter.java
@@ -1,4 +1,4 @@
-package com.microsoft.kiota.serialization.json;
+package com.microsoft.kiota.serialization;
 
 import com.microsoft.kiota.serialization.SerializationWriter;
 import com.microsoft.kiota.serialization.ValuedEnum;
