+++ serialization/java/json/lib/src/main/java/com/microsoft/kiota/serialization/JsonSerializationWriterFactory.java
@@ -1,4 +1,4 @@
-package com.microsoft.kiota.serialization.json;
+package com.microsoft.kiota.serialization;
 
 import java.util.Objects;
