+++ serialization/java/json/lib/src/test/java/com/microsoft/kiota/serialization/LibraryTest.java
@@ -1,7 +1,7 @@
 /*
  * This Java source file was generated by the Gradle 'init' task.
  */
-package com.microsoft.kiota.serialization.json;
+package com.microsoft.kiota.serialization;
 
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
